Move each type of functionality to their own class.
